% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_processing.R
\name{aggregate_imd_by_frequency}
\alias{aggregate_imd_by_frequency}
\title{Aggregate IMD data by temporal frequency}
\usage{
aggregate_imd_by_frequency(data, frequency = "daily")
}
\arguments{
\item{data}{data.frame with IMD data}

\item{frequency}{Character string: "daily", "monthly", "yearly"}
}
\value{
Aggregated data.frame
}
\description{
Aggregate IMD data by temporal frequency
}
