% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{categorize_variables}
\alias{categorize_variables}
\title{Categorize variables by aggregation method}
\usage{
categorize_variables(var_names)
}
\arguments{
\item{var_names}{Character vector of variable names}
}
\value{
Named list with variables categorized by aggregation method
}
\description{
Categorize variables by aggregation method
}
