% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{check_cache}
\alias{check_cache}
\title{Check if data exists in cache}
\usage{
check_cache(cache_key)
}
\arguments{
\item{cache_key}{Cache key to check}
}
\value{
Path to cached file if exists, NULL otherwise
}
\description{
Check if data exists in cache
}
\keyword{internal}
