% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/utils.R
\name{CheckSystemReadiness}
\alias{CheckSystemReadiness}
\alias{check_system_readiness}
\title{Check system readiness for varunayan operations}
\usage{
CheckSystemReadiness(...)

check_system_readiness(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print detailed information}
}
\value{
Logical indicating if system is ready (invisible)
}
\description{
Check system readiness for varunayan operations
}
