% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean up temporary files matching pattern}
\usage{
cleanup_temp_files(pattern = "varunayan_.*", directory = tempdir())
}
\arguments{
\item{pattern}{Regular expression pattern to match files}

\item{directory}{Directory to search (default: tempdir())}
}
\value{
Number of files removed
}
\description{
Clean up temporary files matching pattern
}
