% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{combine_netcdf_files}
\alias{combine_netcdf_files}
\title{Combine multiple NetCDF files into single data.frame}
\usage{
combine_netcdf_files(netcdf_files, verbose = FALSE)
}
\arguments{
\item{netcdf_files}{Character vector of NetCDF file paths}
}
\value{
data.frame with combined data from all files
}
\description{
Combine multiple NetCDF files into single data.frame
}
