% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_geojson}
\alias{convert_to_geojson}
\title{Convert data to GeoJSON format}
\usage{
convert_to_geojson(data)
}
\arguments{
\item{data}{Data to convert (must be data.frame with lon/lat columns)}
}
\value{
GeoJSON formatted data
}
\description{
Convert data to GeoJSON format
}
