% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_geojson_from_bbox}
\alias{create_geojson_from_bbox}
\title{Create GeoJSON from bounding box coordinates}
\usage{
create_geojson_from_bbox(north, south, east, west)
}
\arguments{
\item{north}{Northern latitude}

\item{south}{Southern latitude}

\item{east}{Eastern longitude}

\item{west}{Western longitude}
}
\value{
GeoJSON object representing the bounding box
}
\description{
Create GeoJSON from bounding box coordinates
}
