% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_temp_geojson_file}
\alias{create_temp_geojson_file}
\title{Create temporary GeoJSON file}
\usage{
create_temp_geojson_file(geojson_data, request_id)
}
\arguments{
\item{geojson_data}{GeoJSON data to write}

\item{request_id}{Request identifier for unique filename}
}
\value{
Path to temporary GeoJSON file
}
\description{
Create temporary GeoJSON file
}
