% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/search_and_desc.R
\name{DescribeVariables}
\alias{DescribeVariables}
\alias{describe_variables}
\title{Describe ERA5 variables in detail}
\usage{
DescribeVariables(...)

describe_variables(variables, dataset_type = "all")
}
\arguments{
\item{variables}{Character vector of variable names to describe}

\item{dataset_type}{Character string indicating dataset type ("single", "pressure", or "all")}
}
\value{
data.frame with detailed variable descriptions
}
\description{
Describe ERA5 variables in detail
}
\examples{
\dontrun{
# Describe specific variables
descriptions <- describe_variables(c("2m_temperature", "total_precipitation"))

# Get descriptions for pressure level variables
pressure_desc <- describe_variables("temperature", dataset_type = "pressure")
}
}
