% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.cru_ts_needed_decades}
\alias{.cru_ts_needed_decades}
\title{Find which decade files overlap with a year range}
\usage{
.cru_ts_needed_decades(start_year, end_year)
}
\description{
Find which decade files overlap with a year range
}
\keyword{internal}
