% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.cru_ts_url}
\alias{.cru_ts_url}
\title{Build CRU TS download URL for a specific decade file}
\usage{
.cru_ts_url(variable, decade_start, decade_end)
}
\arguments{
\item{variable}{CRU TS variable code.}

\item{decade_start}{Start year of the decade file.}

\item{decade_end}{End year of the decade file.}
}
\value{
Character URL.
}
\description{
Build CRU TS download URL for a specific decade file
}
\keyword{internal}
