% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.download_cru_ts_decade}
\alias{.download_cru_ts_decade}
\title{Download and decompress a single CRU TS decade file}
\usage{
.download_cru_ts_decade(
  variable,
  decade_start,
  decade_end,
  output_dir,
  use_cache
)
}
\description{
Download and decompress a single CRU TS decade file
}
\keyword{internal}
