% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.expand_hadex3_annual}
\alias{.expand_hadex3_annual}
\title{Expand annual NetCDF array to data.frame}
\usage{
.expand_hadex3_annual(vals, lat, lon, years, fill_val)
}
\description{
Expand annual NetCDF array to data.frame
}
\keyword{internal}
