% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.expand_hadex3_monthly}
\alias{.expand_hadex3_monthly}
\title{Expand monthly NetCDF array to data.frame}
\usage{
.expand_hadex3_monthly(vals, lat, lon, year_months, fill_val)
}
\description{
Expand monthly NetCDF array to data.frame
}
\keyword{internal}
