% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.fetch_hadex3}
\alias{.fetch_hadex3}
\title{Fetch HadEX3 index data (with caching)}
\usage{
.fetch_hadex3(index, frequency, baseline, use_cache, output_dir)
}
\description{
Fetch HadEX3 index data (with caching)
}
\keyword{internal}
