% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.parse_cru_ts_time}
\alias{.parse_cru_ts_time}
\title{Parse CRU TS time coordinate to Date vector}
\usage{
.parse_cru_ts_time(time_var, time_units)
}
\description{
Parse CRU TS time coordinate to Date vector
}
\keyword{internal}
