% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.parse_hadex3_months}
\alias{.parse_hadex3_months}
\title{Parse year+month from monthly HadEX3 time coordinate}
\usage{
.parse_hadex3_months(time_var, time_units)
}
\description{
Parse year+month from monthly HadEX3 time coordinate
}
\keyword{internal}
