% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.parse_hadex3_years}
\alias{.parse_hadex3_years}
\title{Parse year values from annual HadEX3 time coordinate}
\usage{
.parse_hadex3_years(time_var, time_units)
}
\description{
HadEX3 annual files use units "years" (plain integers 1901-2018).
}
\keyword{internal}
