% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.parse_nc_time}
\alias{.parse_nc_time}
\title{Parse numeric time values from NetCDF}
\usage{
.parse_nc_time(time_vals, time_units = NULL)
}
\arguments{
\item{time_vals}{Numeric time values}

\item{time_units}{Character string from the NetCDF time dimension units attribute
(e.g. "seconds since 1970-01-01", "hours since 1900-01-01")}
}
\value{
POSIXct vector
}
\description{
Parse numeric time values from NetCDF
}
\keyword{internal}
