% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.read_cru_ts_nc}
\alias{.read_cru_ts_nc}
\title{Read a CRU TS NetCDF file into a data.frame}
\usage{
.read_cru_ts_nc(nc_file, variable)
}
\arguments{
\item{nc_file}{Path to decompressed .nc file.}

\item{variable}{CRU TS variable code.}
}
\value{
data.frame with year, month, latitude, longitude, value, variable.
}
\description{
Read a CRU TS NetCDF file into a data.frame
}
\keyword{internal}
