% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.read_ncdf4_fast}
\alias{.read_ncdf4_fast}
\title{Fast NetCDF reader using ncdf4 + data.table}
\usage{
.read_ncdf4_fast(netcdf_file)
}
\arguments{
\item{netcdf_file}{Path to NetCDF file}
}
\value{
data.frame in long format (datetime, latitude, longitude, variable, value)
}
\description{
Fast NetCDF reader using ncdf4 + data.table
}
\keyword{internal}
