% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\name{.validate_cru_ts_inputs}
\alias{.validate_cru_ts_inputs}
\title{Validate CRU TS inputs}
\usage{
.validate_cru_ts_inputs(variable, start_year, end_year)
}
\description{
Validate CRU TS inputs
}
\keyword{internal}
