% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{.validate_hadex3_inputs}
\alias{.validate_hadex3_inputs}
\title{Validate HadEX3 inputs}
\usage{
.validate_hadex3_inputs(index, frequency, baseline, start_year, end_year)
}
\description{
Validate HadEX3 inputs
}
\keyword{internal}
