% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.varunayan_env}
\alias{.varunayan_env}
\title{Utility functions for Varunayan R package}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.varunayan_env
}
\description{
Utility functions for Varunayan R package
}
\keyword{datasets}
