% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_output_dir}
\alias{ensure_output_dir}
\title{Create output directory if it doesn't exist}
\usage{
ensure_output_dir(output_dir)
}
\arguments{
\item{output_dir}{Path to output directory}
}
\value{
Character string path to created/existing directory
}
\description{
Create output directory if it doesn't exist
}
