% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\docType{data}
\name{exclude_cols}
\alias{exclude_cols}
\title{Columns to exclude from aggregation}
\format{
An object of class \code{character} of length 10.
}
\usage{
exclude_cols
}
\description{
These are coordinate and metadata columns that should not be aggregated.
}
\keyword{datasets}
