% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_processing.R
\name{filter_imd_by_bbox}
\alias{filter_imd_by_bbox}
\title{Filter IMD data by spatial boundaries}
\usage{
filter_imd_by_bbox(data, north, south, east, west)
}
\arguments{
\item{data}{data.frame with IMD data}

\item{north}{Northern latitude boundary}

\item{south}{Southern latitude boundary}

\item{east}{Eastern longitude boundary}

\item{west}{Western longitude boundary}
}
\value{
Filtered data.frame
}
\description{
Filter IMD data by spatial boundaries
}
