% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_processing.R
\name{filter_imd_by_geojson}
\alias{filter_imd_by_geojson}
\title{Filter IMD data by GeoJSON polygon}
\usage{
filter_imd_by_geojson(data, geojson_file)
}
\arguments{
\item{data}{data.frame with IMD data}

\item{geojson_file}{Path to GeoJSON file}
}
\value{
Filtered data.frame
}
\description{
Filter IMD data by GeoJSON polygon
}
