% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_netcdf_by_bbox}
\alias{filter_netcdf_by_bbox}
\title{Filter NetCDF data by spatial boundaries (bounding box)}
\usage{
filter_netcdf_by_bbox(netcdf_file, north, south, east, west)
}
\arguments{
\item{netcdf_file}{Character string path to NetCDF file}

\item{north}{Northern latitude boundary}

\item{south}{Southern latitude boundary}

\item{east}{Eastern longitude boundary}

\item{west}{Western longitude boundary}
}
\value{
data.frame with spatially filtered data
}
\description{
Filter NetCDF data by spatial boundaries (bounding box)
}
