% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_netcdf_by_geojson}
\alias{filter_netcdf_by_geojson}
\title{Filter NetCDF data by GeoJSON polygon}
\usage{
filter_netcdf_by_geojson(netcdf_file, geojson_file)
}
\arguments{
\item{netcdf_file}{Character string path to NetCDF file}

\item{geojson_file}{Character string path to GeoJSON file}
}
\value{
data.frame with spatially filtered data
}
\description{
Filter NetCDF data by GeoJSON polygon
}
