% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date_range}
\alias{format_date_range}
\title{Format date range for CDS API requests}
\usage{
format_date_range(start_date, end_date)
}
\arguments{
\item{start_date}{Start date (Date object)}

\item{end_date}{End date (Date object)}
}
\value{
Character vector of date strings in YYYY-MM-DD format
}
\description{
Format date range for CDS API requests
}
