% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_pressure_levels}
\alias{format_pressure_levels}
\title{Convert pressure levels to character format required by CDS API}
\usage{
format_pressure_levels(pressure_levels)
}
\arguments{
\item{pressure_levels}{Numeric or character vector of pressure levels}
}
\value{
Character vector of pressure levels
}
\description{
Convert pressure levels to character format required by CDS API
}
