% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_and_desc.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get available dataset types}
\usage{
get_available_datasets()
}
\value{
Character vector of available dataset types
}
\description{
Get available dataset types
}
