% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/search_and_desc.R
\name{GetAvailablePressureLevels}
\alias{GetAvailablePressureLevels}
\alias{get_available_pressure_levels}
\title{Get standard pressure levels available in ERA5}
\usage{
GetAvailablePressureLevels(...)

get_available_pressure_levels()
}
\value{
Character vector of pressure levels in hPa
}
\description{
Get standard pressure levels available in ERA5
}
