% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bbox_from_sf}
\alias{get_bbox_from_sf}
\title{Get bounding box from sf object}
\usage{
get_bbox_from_sf(sf_obj)
}
\arguments{
\item{sf_obj}{sf object}
}
\value{
Named numeric vector with xmin, ymin, xmax, ymax
}
\description{
Get bounding box from sf object
}
