% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/cache.R
\name{GetCacheDir}
\alias{GetCacheDir}
\alias{get_cache_dir}
\title{Cache management for downloaded NetCDF files}
\usage{
GetCacheDir(...)

get_cache_dir(create = TRUE)
}
\arguments{
\item{...}{Arguments passed to the main function (used by aliases).}

\item{create}{Create directory if it doesn't exist (default: TRUE)}
}
\value{
Character string with cache directory path
}
\description{
Functions to cache ERA5 and IMD NetCDF files for faster successive queries
Get cache directory path
}
