% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_pressure_level_variable_names}
\alias{get_pressure_level_variable_names}
\title{Get all valid pressure-level variable names}
\usage{
get_pressure_level_variable_names()
}
\value{
Character vector of valid pressure-level variable names
}
\description{
Get all valid pressure-level variable names
}
