% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_and_desc.R
\name{get_pressure_level_variables}
\alias{get_pressure_level_variables}
\title{Get pressure-level variables dataset}
\usage{
get_pressure_level_variables()
}
\value{
List containing pressure-level variable categories
}
\description{
Get pressure-level variables dataset
}
