% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_single_level_variable_names}
\alias{get_single_level_variable_names}
\title{Get all valid single-level variable names}
\usage{
get_single_level_variable_names()
}
\value{
Character vector of valid single-level variable names
}
\description{
Get all valid single-level variable names
}
