% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_and_desc.R
\name{get_single_level_variables}
\alias{get_single_level_variables}
\title{Get single-level variables dataset}
\usage{
get_single_level_variables()
}
\value{
List containing single-level variable categories
}
\description{
Get single-level variables dataset
}
