% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_values}
\alias{get_time_values}
\title{Get time values based on frequency}
\usage{
get_time_values(frequency)
}
\arguments{
\item{frequency}{Character string indicating temporal frequency}
}
\value{
Character vector of time strings
}
\description{
Get time values based on frequency
}
