% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadex3.R
\name{hadex3}
\alias{hadex3}
\alias{.HADEX3_ANNUAL_INDICES}
\title{HadEX3 Climate Extremes Dataset}
\usage{
.HADEX3_ANNUAL_INDICES
}
\description{
Functions to download and process the Met Office HadEX3 global climate
extremes dataset based on ETCCDI indices.
}
\references{
Dunn et al. (2020) Development of an updated global land in-situ-based dataset
of temperature and precipitation extremes: HadEX3. J. Geophys. Res. Atmos.
https://doi.org/10.1029/2019JD032263
}
