% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{heat_index}
\alias{heat_index}
\title{Calculate heat index (apparent temperature)}
\usage{
heat_index(temp_c, rh)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{rh}{Relative humidity as percentage (0-100)}
}
\value{
Heat index in celsius
}
\description{
Calculate heat index (apparent temperature)
}
\examples{
heat_index(35, 70)
}
\references{
Rothfusz, L.P. (1990). The Heat Index Equation.
NWS Southern Region Technical Attachment, SR/SSD 90-23.
https://www.weather.gov/media/ffc/ta_htindx.PDF
}
