% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{heat_index_from_dewpoint}
\alias{heat_index_from_dewpoint}
\title{Calculate heat index from dewpoint}
\usage{
heat_index_from_dewpoint(temp_c, dewpoint_c)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{dewpoint_c}{Dewpoint temperature in celsius}
}
\value{
Heat index in celsius
}
\description{
Calculate heat index from dewpoint
}
\examples{
heat_index_from_dewpoint(35, 25)
}
