% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{heat_index_risk_category}
\alias{heat_index_risk_category}
\title{Heat index risk categories}
\usage{
heat_index_risk_category(hi)
}
\arguments{
\item{hi}{Heat index value in Celsius}
}
\value{
Factor with risk category
}
\description{
Classify heat stress risk based on heat index using NWS categories.
}
\examples{
heat_index_risk_category(c(27, 33, 40, 46, 55))
}
