% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{humidex}
\alias{humidex}
\title{Calculate humidex (Canadian heat index)}
\usage{
humidex(temp_c, dewpoint_c)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{dewpoint_c}{Dewpoint temperature in celsius}
}
\value{
Humidex in celsius
}
\description{
Calculate humidex (Canadian heat index)
}
\examples{
humidex(30, 22)
}
\references{
Masterson, J. & Richardson, F.A. (1979). Humidex, A Method of
Quantifying Human Discomfort Due to Excessive Heat and Humidity.
Environment Canada, Atmospheric Environment Service.
}
