% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{is_max_var}
\alias{is_max_var}
\title{Check if a variable should use maximum}
\usage{
is_max_var(var_name)
}
\arguments{
\item{var_name}{Character string variable name}
}
\value{
Logical indicating if variable should use max
}
\description{
Check if a variable should use maximum
}
