% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{is_min_var}
\alias{is_min_var}
\title{Check if a variable should use minimum}
\usage{
is_min_var(var_name)
}
\arguments{
\item{var_name}{Character string variable name}
}
\value{
Logical indicating if variable should use min
}
\description{
Check if a variable should use minimum
}
