% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{is_rate_var}
\alias{is_rate_var}
\title{Check if a variable is a rate variable}
\usage{
is_rate_var(var_name)
}
\arguments{
\item{var_name}{Character string variable name}
}
\value{
Logical indicating if variable is a rate
}
\description{
Check if a variable is a rate variable
}
