% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{is_sum_var}
\alias{is_sum_var}
\title{Check if a variable should be summed}
\usage{
is_sum_var(var_name)
}
\arguments{
\item{var_name}{Character string variable name}
}
\value{
Logical indicating if variable should be summed
}
\description{
Check if a variable should be summed
}
