% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_geojson}
\alias{is_valid_geojson}
\title{Check if data is valid GeoJSON}
\usage{
is_valid_geojson(data)
}
\arguments{
\item{data}{List or data structure to check}
}
\value{
Logical indicating if data is valid GeoJSON
}
\description{
Check if data is valid GeoJSON
}
