% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_download.R
\name{list_imd_datasets}
\alias{list_imd_datasets}
\title{List available IMD datasets}
\usage{
list_imd_datasets()
}
\value{
data.frame with available IMD datasets and their characteristics
}
\description{
List available IMD datasets
}
