% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\docType{data}
\name{max_vars}
\alias{max_vars}
\title{Variables that should use maximum aggregation}
\format{
An object of class \code{character} of length 8.
}
\usage{
max_vars
}
\description{
These variables represent maximum values that should use max() when aggregating.
}
\keyword{datasets}
