% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\docType{data}
\name{min_vars}
\alias{min_vars}
\title{Variables that should use minimum aggregation}
\format{
An object of class \code{character} of length 4.
}
\usage{
min_vars
}
\description{
These variables represent minimum values that should use min() when aggregating.
}
\keyword{datasets}
