% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_date_flexible}
\alias{parse_date_flexible}
\title{Parse date string flexibly}
\usage{
parse_date_flexible(date_string)
}
\arguments{
\item{date_string}{Character string in various date formats}
}
\value{
Date object
}
\description{
Parse date string flexibly
}
