% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\docType{data}
\name{rate_vars}
\alias{rate_vars}
\title{Rate variables that should be averaged}
\format{
An object of class \code{character} of length 84.
}
\usage{
rate_vars
}
\description{
These are rate variables that should be averaged (not summed) when aggregating
to longer time periods.
}
\keyword{datasets}
