% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_processing.R
\name{read_imd_rainfall_ncdf4}
\alias{read_imd_rainfall_ncdf4}
\title{Read IMD NetCDF using ncdf4}
\usage{
read_imd_rainfall_ncdf4(file_path, grid_specs, year)
}
\arguments{
\item{file_path}{Path to NetCDF file}

\item{grid_specs}{Grid specifications}

\item{year}{Year of data}
}
\value{
data.frame
}
\description{
Read IMD NetCDF using ncdf4
}
\keyword{internal}
