% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{relative_humidity_from_dewpoint}
\alias{relative_humidity_from_dewpoint}
\title{Calculate relative humidity from temperature and dewpoint}
\usage{
relative_humidity_from_dewpoint(temp_c, dewpoint_c)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{dewpoint_c}{Dewpoint temperature in celsius}
}
\value{
Relative humidity as percentage (0-100)
}
\description{
Uses the Magnus-Tetens approximation to calculate relative humidity
from air temperature and dewpoint temperature.
}
\examples{
relative_humidity_from_dewpoint(25, 18)
}
