% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save file to cache}
\usage{
save_to_cache(file_path, cache_key)
}
\arguments{
\item{file_path}{Path to file to cache}

\item{cache_key}{Cache key for the file}
}
\value{
Path to cached file
}
\description{
Save file to cache
}
\keyword{internal}
