% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/cache.R
\name{ShowCacheInfo}
\alias{ShowCacheInfo}
\alias{show_cache_info}
\title{Show cache contents and statistics}
\usage{
ShowCacheInfo(...)

show_cache_info()
}
\arguments{
\item{...}{Arguments passed to the main function (used by aliases).}
}
\value{
data.frame with cache information (invisible)
}
\description{
Show cache contents and statistics
}
