% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_operation}
\alias{time_operation}
\title{Time an operation and log the duration}
\usage{
time_operation(expr, label, verbose = TRUE)
}
\arguments{
\item{expr}{Expression to time}

\item{label}{Label for the operation}

\item{verbose}{Whether to print timing info (default: TRUE)}
}
\value{
Result of the expression
}
\description{
Time an operation and log the duration
}
