% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{utci_category}
\alias{utci_category}
\title{UTCI thermal stress categories}
\usage{
utci_category(utci_val)
}
\arguments{
\item{utci_val}{UTCI value in Celsius}
}
\value{
Factor with thermal stress category
}
\description{
Classify thermal stress based on UTCI values.
}
\examples{
utci_category(c(-20, 0, 20, 30, 40))
}
