% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_bbox}
\alias{validate_bbox}
\title{Validate bounding box coordinates}
\usage{
validate_bbox(north, south, east, west)
}
\arguments{
\item{north}{Northern boundary}

\item{south}{Southern boundary}

\item{east}{Eastern boundary}

\item{west}{Western boundary}
}
\value{
Logical indicating if bounding box is valid
}
\description{
Validate bounding box coordinates
}
