% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_coordinates}
\alias{validate_coordinates}
\title{Validate geographic coordinates}
\usage{
validate_coordinates(lat, lon)
}
\arguments{
\item{lat}{Latitude value}

\item{lon}{Longitude value}
}
\value{
Logical indicating if coordinates are valid
}
\description{
Validate geographic coordinates
}
