% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{vapor_pressure_from_dewpoint}
\alias{vapor_pressure_from_dewpoint}
\title{Calculate vapor pressure from dewpoint temperature}
\usage{
vapor_pressure_from_dewpoint(dewpoint_c)
}
\arguments{
\item{dewpoint_c}{Dewpoint temperature in celsius}
}
\value{
Vapor pressure in hPa (hectopascals)
}
\description{
Calculate vapor pressure from dewpoint temperature
}
\examples{
vapor_pressure_from_dewpoint(20)
}
