% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{wbgt_risk_category}
\alias{wbgt_risk_category}
\title{heat stress risk categories}
\usage{
wbgt_risk_category(wbgt)
}
\arguments{
\item{wbgt}{WBGT value in celsius}
}
\value{
Factor with risk category
}
\description{
Classify heat stress risk based on WBGT values using ISO 7243 guidelines.
}
\examples{
wbgt_risk_category(c(25, 28, 31, 34))
}
