% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{wet_bulb_from_dewpoint}
\alias{wet_bulb_from_dewpoint}
\title{Calculate wet bulb temperature from dewpoint}
\usage{
wet_bulb_from_dewpoint(temp_c, dewpoint_c)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{dewpoint_c}{Dewpoint temperature in celsius}
}
\value{
Wet bulb temperature in celsius
}
\description{
Calculate wet bulb temperature from dewpoint
}
\examples{
wet_bulb_from_dewpoint(30, 22)
}
