% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{wet_bulb_temperature}
\alias{wet_bulb_temperature}
\title{Calculate wet bulb temperature}
\usage{
wet_bulb_temperature(temp_c, rh)
}
\arguments{
\item{temp_c}{Air temperature in celsius}

\item{rh}{Relative humidity as percentage (0-100)}
}
\value{
Wet bulb temperature in Celsius
}
\description{
Calculates wet bulb temperature using the Stull (2011) empirical formula.
Valid for temperatures between -20C and 50C and RH between 5\% and 99\%.
}
\examples{
wet_bulb_temperature(30, 60)
}
\references{
Stull, R. (2011). Wet-Bulb Temperature from Relative Humidity
and Air Temperature. Journal of Applied Meteorology and Climatology,
50(11), 2267-2269. https://journals.ametsoc.org/view/journals/apme/50/11/jamc-d-11-0143.1.xml
}
