% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{aggregate_by_frequency}
\alias{aggregate_by_frequency}
\title{Aggregate data by temporal frequency}
\usage{
aggregate_by_frequency(data, frequency, verbose = FALSE)
}
\arguments{
\item{data}{data.frame with datetime column}

\item{frequency}{Character string: "daily", "monthly", "yearly"}

\item{verbose}{Logical indicating whether to print progress messages}
}
\value{
data.frame with temporally aggregated data
}
\description{
Aggregate data by temporal frequency
}
