% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cru_ts.R
\docType{data}
\name{cru_ts}
\alias{cru_ts}
\alias{.CRU_TS_VERSION}
\title{CRU TS Gridded Climate Dataset}
\format{
An object of class \code{character} of length 1.
}
\usage{
.CRU_TS_VERSION
}
\description{
Functions to download and process the University of East Anglia Climatic
Research Unit Time-Series (CRU TS) gridded monthly climate dataset.
Covers global land areas (excl. Antarctica) at 0.5 degree resolution,
1901-2022.
}
\references{
Harris et al. (2020) Version 4 of the CRU TS monthly high-resolution
gridded multivariate climate dataset. Sci. Data 7:109.
https://doi.org/10.1038/s41597-020-0453-3
}
\keyword{datasets}
