% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/utils.R
\name{GetPackageInfo}
\alias{GetPackageInfo}
\alias{get_package_info}
\title{Get package version information}
\usage{
GetPackageInfo(...)

get_package_info()
}
\arguments{
\item{...}{Arguments passed to the main function (used by aliases).}
}
\value{
data.frame with package versions (invisible)
}
\description{
Get package version information
}
