% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R, R/search_and_desc.R
\name{ListAvailableVariables}
\alias{ListAvailableVariables}
\alias{list_available_variables}
\title{List all available variables for a dataset type}
\usage{
ListAvailableVariables(...)

list_available_variables(dataset_type = "all")
}
\arguments{
\item{...}{Arguments passed to the main function (used by aliases).}

\item{dataset_type}{Character string indicating dataset type}
}
\value{
Character vector of variable names
}
\description{
List all available variables for a dataset type
}
