% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{rate_vars}
\alias{rate_vars}
\title{Rate variables that should be averaged}
\usage{
rate_vars
}
\description{
These are rate variables that should be averaged (not summed) when aggregating
to longer time periods.
}
