% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{sum_vars}
\alias{sum_vars}
\title{Variables that should be summed over time}
\usage{
sum_vars
}
\description{
Accumulation variables (precipitation, radiation) that use sum for temporal
aggregation.
}
