% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lists.R
\name{sum_vars}
\alias{sum_vars}
\title{Variables that should be summed over time}
\usage{
sum_vars
}
\description{
These are accumulation variables like precipitation and radiation that should
be summed when aggregating to longer time periods.
}
