% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varunayan-package.R
\docType{package}
\name{varunayan-package}
\alias{varunayan}
\alias{varunayan-package}
\title{varunayan: Package for downloading and processing climatic variables from ERA5 and IMD}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package for downloading and processing ERA5 climate data, IMD (Indian Meteorological Department) gridded data, and HadEX3 global climate extremes (ETCCDI indices) for custom geographical regions. Provides functions to extract climate data using bounding boxes, GeoJSON polygons, or point coordinates and calculate heat indices. Supports various temporal frequencies and spatial resolutions with analysis-ready outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saketlab.github.io/varunayanR}
  \item \url{https://github.com/saketlab/varunayanR}
  \item Report bugs at \url{https://github.com/saketlab/varunayanR/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

Authors:
\itemize{
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})
}

}
\keyword{internal}
